==> 3_Mortality/1_NCDLRI.ipynb <==
import json
import salem
import math
import numpy as np
import pandas as pd
import geopandas as gpd

with open("../1_Country_Emission/Countries.json",'r') as load_f:
    Countries = json.load(load_f)

NCDLRI = pd.read_csv('C:/Users/15311/Desktop/InMAP/Mortality Number/NCD_LRI.csv')
code = pd.read_csv('C:/Users/15311/Desktop/InMAP/Mortality Number/code.csv')

NCDLRI = NCDLRI[['year','location_name','sex_name','age_name','cause_name','metric_name','val']]
NCDLRI = NCDLRI[NCDLRI['metric_name']  == 'Rate']

LRI = NCDLRI[NCDLRI['cause_name'] == 'Lower respiratory infections'].reset_index(drop=True)
NCD = NCDLRI[NCDLRI['cause_name'] == 'Non-communicable diseases'].reset_index(drop=True)

code.columns = ['location_name','code']

left_countries = []
for i in NCD['location_name']:
    if i not in list(code['location_name']):
        left_countries.append(i)

There are some countries in GBD dataset but not in countries' codes dataset.

left_countries

NCD_left = 0
LRI_left = 0
for i in left_countries:
    NCD_left += NCD[NCD['location_name'] == i]['val'].values[0]
    LRI_left += LRI[LRI['location_name'] == i]['val'].values[0]

NCD_left,LRI_left # Add those number of death to "ROW"

LRI = pd.merge(LRI,code)
NCD = pd.merge(NCD,code)

NCD

missing_countries = []
for i in Countries:
    if i in NCD['code'].unique():
        continue    
    else:
        missing_countries.append(i)

missing_countries

# choose a region has similar GPD and in the same region.
# or: Assign those countries according to world bank region

df_shp = salem.read_shapefile(r'C:/Users/15311/Desktop/natural_earth_vector/10m_cultural/ne_10m_admin_0_countries.shp')

df_shp[df_shp['ADM0_A3'].isin(['PSX','SDN'])]

df_shp.loc[11,'ADM0_A3'], df_shp.loc[86,'ADM0_A3']

### Change 'PSX' to 'PSE' and change 'SDN' to  'SUD'

df_shp.loc[11,'ADM0_A3'] = 'PSE'
df_shp.loc[86,'ADM0_A3'] = 'SUD'

df_shp.loc[11,'ADM0_A3'], df_shp.loc[86,'ADM0_A3']

subset = df_shp[df_shp["ADM0_A3"].isin(missing_countries[:-1])]

# subset.columns # ‘REGION_WB’

for i in missing_countries:
    if i in subset['ADM0_A3'].unique():
        continue
    else:
        print(i)

# np.sort(list(NCD['location_name'].unique()))

for i in range(len(LRI)):
    if LRI.iloc[i,7] not in Countries:
        LRI.iloc[i,7] = 'ROW' ## No.7 is the country code
        
for i in range(len(NCD)):
    if NCD.iloc[i,7] not in Countries:
        NCD.iloc[i,7] = 'ROW'

NCD_dict = {}
LRI_dict = {}
for i in Countries:
    NCD_dict[i] = 0
    LRI_dict[i] = 0

GBD_WB = pd.read_csv('C:/Users/15311/Desktop/InMAP/Mortality Number/GBD_WB.csv')

GBD_LRI = GBD_WB[GBD_WB['cause_name'] == 'Lower respiratory infections']
GBD_NCD = GBD_WB[GBD_WB['cause_name'] == 'Non-communicable diseases']

GBD_NCD

NCD_dict = {}
LRI_dict = {}
count = 0
for i in Countries[:-1]:
    if i in list(subset[subset['REGION_WB'] == 'Latin America & Caribbean']['ADM0_A3']):
        NCD_dict[i] = GBD_NCD[GBD_NCD['location_name'] == 'Latin America & Caribbean']['val'].values[0]
        LRI_dict[i] = GBD_LRI[GBD_LRI['location_name'] == 'Latin America & Caribbean']['val'].values[0]

    if i in list(subset[subset['REGION_WB'] == 'Middle East & North Africa']['ADM0_A3']):
        NCD_dict[i] = GBD_NCD[GBD_NCD['location_name'] == 'Middle East & North Africa']['val'].values[0]
        LRI_dict[i] = GBD_LRI[GBD_LRI['location_name'] == 'Middle East & North Africa']['val'].values[0]
        
    if i in list(subset[subset['REGION_WB'] == 'Sub-Saharan Africa']['ADM0_A3']):
        NCD_dict[i] = GBD_NCD[GBD_NCD['location_name'] == 'Sub-Saharan Africa']['val'].values[0]
        LRI_dict[i] = GBD_LRI[GBD_LRI['location_name'] == 'Sub-Saharan Africa']['val'].values[0]
        
    if i in list(subset[subset['REGION_WB'] == 'Europe & Central Asia']['ADM0_A3']):
        NCD_dict[i] = GBD_NCD[GBD_NCD['location_name'] == 'Europe & Central Asia']['val'].values[0]
        LRI_dict[i] = GBD_LRI[GBD_LRI['location_name'] == 'Europe & Central Asia']['val'].values[0]
        
    if i in list(subset[subset['REGION_WB'] == 'East Asia & Pacific']['ADM0_A3']):
        NCD_dict[i] = GBD_NCD[GBD_NCD['location_name'] == 'East Asia & Pacific']['val'].values[0]
        LRI_dict[i] = GBD_LRI[GBD_LRI['location_name'] == 'East Asia & Pacific']['val'].values[0]
        
    if i not in missing_countries:
        NCD_dict[i] = NCD[NCD['code'] == i]['val'].values[0]
        LRI_dict[i] = LRI[LRI['code'] == i]['val'].values[0]

NCD_dict['ROW'] = sum(NCD[NCD['code'] == 'ROW']['val']) + NCD_left
LRI_dict['ROW'] = sum(LRI[LRI['code'] == 'ROW']['val']) + LRI_left

subset[['SOVEREIGNT','ADM0_A3','REGION_WB']]

NCDLRI_Total = {}
for i in NCD_dict:
    NCDLRI_Total[i] = LRI_dict[i] + NCD_dict[i]

len(NCDLRI_Total),len(LRI_dict),len(NCD_dict) # Without ANT and USR

NCDLRI_Total

NCDLRI_Total['CHN']

NCDLRI_Total['HKG'],NCDLRI_Total['MAC']

# Let Hong Kong and Macao have the same mortality rate as mainland China
NCDLRI_Total['HKG'],NCDLRI_Total['MAC'] = NCDLRI_Total['CHN'], NCDLRI_Total['CHN']

NCDLRI_Total['HKG'],NCDLRI_Total['MAC']

sorted(NCDLRI_Total.items(), key=lambda x: x[1], reverse=True)

file=open("NCDLRI_Total.json",'w')
json.dump(NCDLRI_Total,file)
file.close()
