==> 3_Mortality/2_CountryClassify.ipynb <==
import numpy as np
import pandas as pd
import geopandas as gpd
import json

with open("../2_Concentration_Factors/total_output_sector.json",'r') as load_f: # ../2_Concentration_Factors/
    total_output = json.load(load_f)
Country = list(total_output.keys())

Country.remove('USR')
Country.remove('ANT')

'SUD' in Country,'PSE' in Country

# df_shp = gpd.read_file('10m_cultural/ne_10m_admin_0_countries.shp')
df_shp = gpd.read_file('C:/Users/15311/Desktop/natural_earth_vector/10m_cultural/ne_10m_admin_0_countries.shp')

# Change 'PSX' to 'PSE' and change 'SDN' to  'SUD'
df_shp.loc[11,'ADM0_A3'] = 'PSE'
df_shp.loc[86,'ADM0_A3'] = 'SUD'

# wholeworldB = gpd.read_file('output/output_wholeworld/wholeworld_bio.shp')
wholeworldB = gpd.read_file('C:/Users/15311/Desktop/output_wholeworld/wholeworld_bio.shp')
wholeworldB = gpd.GeoDataFrame(wholeworldB)

wholeworldB['area_fraction'] = [{} for i in range(len(wholeworldB))]

wholeworldB.iloc[1,13]

for i in Country[:-1]:
    for j in range(len(wholeworldB)):
        a = df_shp[df_shp['ADM0_A3']==i].intersection(wholeworldB.iloc[j].geometry).area.values[0]
        if a == 0:
            pass
        else:
            wholeworldB.iloc[j,13][i] = a / wholeworldB.iloc[j].geometry.area

wholeworldB

# Summation of the value insde a dictionary
def dictsum(myDict):      
    sum = 0
    for i in myDict: 
        sum = sum + myDict[i] 
    return sum

for i in range(len(wholeworldB)):
    if wholeworldB.iloc[i,13] == {}:
        wholeworldB.iloc[i,13]['ROW'] = 1
    elif dictsum(wholeworldB.iloc[i,13]) != 1:
        wholeworldB.iloc[i,13]['ROW'] = 1 - dictsum(wholeworldB.iloc[i,13])

wholeworldB.to_csv('comparison/country_classify.csv')
