==> 3_Mortality/4_MortPerConcentration.ipynb <==
import pandas as pd
import numpy as np
import geopandas as gpd
import salem
import copy
import json

def hazard_ratio(theta, alpha, miu, v, concentration):
    z = (concentration-2.4) if (concentration - 2.4) > 0 else 0
    numerator = theta * np.log(z/(alpha)+1)
    denominator = 1 + np.exp(-(z-miu)/v)
    GEMM = np.exp(numerator/denominator)
    return GEMM 
def hazard_ratio_NCDLRI(concentration):
    if concentration < 2.4: # Lowest observed concentration in any of the 41 cohorts
        return 1
    else:
        return hazard_ratio(0.1430, 1.6, 15.5, 36.8, concentration)

fraction = pd.read_csv('comparison/country_classify.csv')
# fraction = pd.read_csv('C:/Users/15311/Desktop/country_classify.csv')

# file = open("comparison/Baseline.json", "r")
file = open("C:/Users/15311/Desktop/Baseline.json", "r")
baseline = file.read()
baseline = json.loads(baseline)
Country = list(baseline.keys())

## Calculate the total population of each country

file = open("comparison/Total_pop.json", "r")
# file = open("C:/Users/15311/Desktop/Total_pop_old.json", "r")
Total_pop = file.read()
Total_pop = json.loads(Total_pop)

## Allocate baseline mortality rate to each grid cell (Submit to cc part 1)

# For ROW
A = ['agr','ene','ind','nrtr','rcor','rcoc','rcoo','shp','slv','wst']
for j in A:
    read_shp = gpd.read_file(r'output/output_wholeworld/wholeworld_' + j + '.' + 'shp')
    fraction['TotalPM25'] += read_shp['TotalPM25']

fraction['base_mort'] = [{} for i in range(len(fraction))]

fraction

for index in range(len(fraction)):
    for country in Country:
        if country not in json.loads(fraction.iloc[index]['area_fraction'].replace('\'','\"')).keys():
            continue
        frac = json.loads(fraction.iloc[index]['area_fraction'].replace('\'','\"'))[country]
        fraction.iloc[index]['base_mort'][country] = frac * fraction.iloc[index]['TotalPop'] / Total_pop[country] * baseline[country]

fraction.to_csv('comparison/baseline_fraction_Hbar1.csv')

## Mortality rate per concentration comparison (Submit to cc part 2)

fraction = pd.read_csv('comparison/baseline_fraction_Hbar1.csv')

fraction['mort_per_con'] = [{} for i in range(len(fraction))]

for i in range(len(fraction)):
    for country in Country:
        if country not in json.loads(fraction.iloc[i]['base_mort'].replace('\'','\"')).keys():
            continue
        ratio = 1 - 1 / hazard_ratio_NCDLRI(fraction.iloc[i]['TotalPM25'])
        fraction.iloc[i]['mort_per_con'][country] = ratio * json.loads(fraction.iloc[i]['base_mort'].replace('\'','\"'))[country] / fraction.iloc[i]['TotalPM25']    

fraction.to_csv('comparison/baseline_fraction_Hbar2.csv')
