==> 3_Mortality/5_concentrations_sectors.ipynb <==
import pandas as pd
import numpy as np
import json
import geopandas as gpd

with open("total_output_sector.json",'r') as load_f:
    total_output = json.load(load_f)
Country = list(total_output.keys())

A = ['agr','ene','ind','nrtr','rcor','rcoc','rcoo','shp','slv','wst']

column_name = []
for i in Country:
    for j in A:
        column_name.append(i+'_'+j)

shp1 = np.zeros(shape=(273739,))
for i in Country:
    for j in A:
        shp = gpd.read_file(r'output' + '/' + 'output' + '_' + i + '/' + i + '_'+j + '.' + 'shp')
        shp1 = np.row_stack((shp1,shp['TotalPM25'].values))

concentration_by_sectors = pd.DataFrame(shp1[1:].T)

concentration_by_sectors.columns = column_name
concentration_by_sectors

concentration_by_sectors.to_csv('comparison/concentration_by_sectors.csv')
