==> 3_Mortality/8_mortality_sectors.ipynb <==
import pandas as pd
import numpy as np
import geopandas as gpd
import salem
import copy
import json
import sys

with open("../2_Concentration_Factors/total_output_sector.json",'r') as load_f:
    total_output = json.load(load_f)
Country = list(total_output.keys())

concentration_matrix = pd.read_csv('C:/Users/15311/Desktop/concentration_by_sectors.csv',index_col=0)
Country_sector = list(concentration_matrix.columns)

## Mortality factor

concentration_matrix

concentration_matrix.to_numpy()

Sub_Country1 = Country_sector[int(sys.argv[1]):int(sys.argv[2])] 
# Sub_Country2 = Country_sector[30:60]
# Sub_Country3 = Country_sector[60:90]
# Sub_Country4 = Country_sector[90:120]
# Sub_Country5 = Country_sector[80:100]
# Sub_Country6 = Country_sector[120:150]
# Sub_Country7 = Country_sector[150:180]
# Sub_Country8 = Country_sector[180:210]
# Sub_Country9 = Country_sector[210:230]

# fraction = pd.read_csv('comparison/baseline_fraction_Hbar2.csv')
fraction = pd.read_csv('C:/Users/15311/Desktop/baseline_fraction_Hbar2.csv')

concentration_matrix['area_fraction'] = fraction['area_fraction']
concentration_matrix['base_mort'] = fraction['base_mort']
concentration_matrix['mort_per_con'] = fraction['mort_per_con']

total_mortality = {}

from tqdm import tqdm
for consume_country in tqdm(Sub_Country1):
    mortality0 = {}
    for country in Country:
        mortality0[country] = 0
        popdeath = 0
        for index in range(len(concentration_matrix)):
            if country not in json.loads(concentration_matrix.iloc[index]['mort_per_con'].replace('\'','\"')).keys():
                continue
            popdeath += concentration_matrix.iloc[index][consume_country] * json.loads(concentration_matrix.iloc[index]['mort_per_con'].replace('\'','\"'))[country]
        mortality0[country] += popdeath
    total_mortality[consume_country] = mortality0

file=open("comparison/Excess_mort_sector"+ str(sys.argv[3]) + ".json",'w')
json.dump(total_mortality,file)
file.close()

list_20 = list(np.arange(0,1900,20))

list_4 = list(np.arange(0,1900,4))

for i in list_4[:-1]:
    for j in list_4[1:]:
        if j - i == 4:
            print(i,j,list_4.index(j))

for i in list_20[:-1]:
    for j in list_20[1:]:
        if j - i == 20:
            print(i,j,list_20.index(j))

# check correctness

with open("C:/Users/15311/Desktop/Excess_mort_sector1.json",'r') as load_f:
    Excess_mort_sector1 = json.load(load_f)

with open("C:/Users/15311/Desktop/Excess_mort_sector2.json",'r') as load_f:
    Excess_mort_sector2 = json.load(load_f)

Excess_mort_sector1.keys()

Excess_mort_sector2.keys()
